NXOS interfaces poller: another xtraction due to os bug
We breakup the way we pull the information off of NXOS to keep the
CPU utilization low, processing time low and the bandwidth consumed
low. The use of show ethernet regex to get the MTU and MAC address
was driven by the behavior on NXOSv where show interfaces took too
long in the presence of SVIs. Hardware boxes don't have this problem.
Plus show ethernet regex has an NXOS bug upto 9.3.4 where the command
fails if there are more than 64 interfaces on the box.

All this leads to using show int instead of show ethernet regex.
Additional missing data or incorrect data such as speed unit fixed as
well.