Add support for saving running config of devices
This saves the running config as is. We strip off any
data such as timestamp that can cause us to keep saving
copies. We poll every hour for this data. Unlike other
tables, this is one table for which the show command returns
exactly the config and not a dataframe.

unique() method is not useful for this config as of now.
summarize() returns only the count of devices for which we have
a config.

This patch includes tests.